% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrhenius.R
\name{arrhenius}
\alias{arrhenius}
\title{Compute the arrhenius version of a soil respiration}
\usage{
arrhenius(param, data_in)
}
\arguments{
\item{param}{input parameter vector - is a data frame with the following columns:
1. name
2. changeable
3. value
4. minVal
5. maxVal}
}
\value{
rSoil at each of the Plots with a given value
}
\description{
\code{arrhenius} returns the predicted soil respiration vlaues
}
