% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize}
\alias{initialize}
\title{Return a random initial start of a parameter set}
\usage{
initialize(param)
}
\arguments{
\item{param}{input parameter vector - is a data frame with the following columns:
1. name
2. changeable
3. value
4. minVal
5.}
}
\value{
updated parameter values, with a "knob" column added for simulating the temperatures.
}
\description{
\code{initialize} Returns a random start value for a parameter set
}
