% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmc_treatment.R
\name{mcmc_treatment}
\alias{mcmc_treatment}
\title{Does a MCMC estimate across different data treatements for a given model}
\usage{
mcmc_treatment(model_id, mcmc_mode, treatment_start = 1)
}
\arguments{
\item{model_id}{The name of the model we are using.  Must be the same name as}

\item{mcmc_mode}{The type of MCMC estimate we want to run}

\item{treatment_start}{Where we are starting our treatment.  Helpful if we restart and stop runs.}
}
\value{
Nothing - this just saves a bunch of files
}
\description{
\code{mcmc_treatment} does the heavy lifting for a given model, splitting the data into treatments, runs and saves an MCMC parameter estimate for a given mode.
}
