% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_chain.R
\name{metropolis_chain}
\alias{metropolis_chain}
\title{Do a metropolis chain}
\usage{
metropolis_chain(param_in, data_in, model, join_key, iterations, tuning,
  estimate_chain, mcmc_parameters)
}
\arguments{
\item{param_in}{input parameter vector - is a data frame with the following columns:
1. name
2. changeable
3. value
4. minVal
5. knob
6. sampled}

\item{data_in}{input data we use to compare in the likelihood function.}

\item{model}{the particular model we are using in the mcmc}

\item{join_key}{The name of the variable we will join to data_in when doing likelihood}

\item{iterations}{The number of times we want to iterate through this chain}

\item{tuning}{a T/F value depending on if we want to limit the searchable range of parameters}

\item{estimate_chain}{a T/F value depending on what we want to return.}

\item{mcmc_parameters}{a data frame of how we want to estimate (can be mcmc_superfast or mcmc_defaults)}
}
\value{
a list, depending on if we want to track parameter values and if they were accepted, or just the final value for updating.
}
\description{
\code{metropolis_chain} Runs a MCMC metropolis chain
}
