% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbes.R
\name{microbes}
\alias{microbes}
\title{Compute the dead_soil version of a model}
\usage{
microbes(param, data_in)
}
\arguments{
\item{param}{input parameter vector - is a data frame with the following columns:
1. name
2. changeable
3. value
4. minVal
5. maxVal}
}
\value{
rSoil at each of the Plots with a given value
}
\description{
\code{microbes} Returns the prior probability of a parameter set
}
