% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_parameters.R
\name{sample_parameters}
\alias{sample_parameters}
\title{Return a random sample of a parameter set}
\usage{
sample_parameters(param)
}
\arguments{
\item{param}{input parameter vector - is a data frame with the following columns:
1. name
2. changeable
3. value
4. minVal
5. knob
6. sampled}
}
\value{
updated parameter values, with a "knob" column added for simulating the temperatures.
}
\description{
\code{sample_parameters} Returns a parameter set where a given parameter has been changed for a parameter set
}
